Remove resizing when too sparse, fix #rehash/#set
Removed resizing when LF reached 0.25 because it was causing some very
strange behavior where the hash_map would bounce around between emptyish
and fullish after a resize. The spec doesn't call for this feature; I
just read about it on Wikipedia and thought I'd try to implement it

hash_map.load is not incremented on rehashes; added a boolean arg to set
to check if it is a rehash, since the values need to be reset *after*
the slots are added to coll.